requirements.txt
torch
torchvision
datasets
scikit-learn
pandas
matplotlib
Pillow
opencv-python
onnx
onnxruntime
tqdm
